+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Improved SuiTransaction `execute` performance
+
 ### Removed
 
 ## [0.24.0] - 2023-06-01